Update logic to only delete resources created by this module
- Use generated uuid to ensure the resource is created by this module instance
- Skip creating reserved namespaces that start with openshift-
- Move test logic to example/ folder

+++ scripts/deleteNamespace.sh
@@ -7,10 +7,27 @@ if [[ -n "${BIN_DIR}" ]]; then
 fi
 
 NAMESPACE="$1"
+UUID="$2"
 
 if [[ -z "${NAMESPACE}" ]]; then
     echo "Namespace is required as the first parameter"
     exit 1
+fi
+
+if ! kubectl get configmap ns-create -n "${NAMESPACE}" 1> /dev/null 2> /dev/null; then
+  echo "Namespace create configmap not found. Skipping destroy"
+  exit 0
+fi
+
+## Check that this module created the namespace before destroying it
+
+OLM_CONFIG=$(kubectl get configmap olm-install -n default -o json | jq -c '.data')
+
+CLUSTER_UUID=$(echo "${OLM_CONFIG}" | jq -r '.uuid // "empty"')
+
+if [[ "${CLUSTER_UUID}" != "${UUID}" ]]; then
+  echo "UUID of module does not match created namespace. Skipping destroy"
+  exit 0
 fi
 
 kubectl delete namespace "${NAMESPACE}" --timeout=5m --ignore-not-found=true
